requirements.txt
django
django-background-tasks
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
drf-yasg
# MODEL DEPENDENCY
google-generativeai
sympy
antlr4-python3-runtime==4.11 # sympy dependency for latex parsing
